chore: document location refresh threshold
Agent-Logs-Url: https://github.com/icebear74/Bewaesserung/sessions/41d5189f-c0ca-4f5c-abfd-eae8989b0f19

+++ Pcf857xDevice.cpp
@@ -23,7 +23,7 @@ bool Pcf857xDevice::digitalWrite(uint8_t channel, bool level) {
     if (level) {
         _state |= mask;
     } else {
-        _state &= (uint16_t)~mask;
+        _state &= ~mask;
     }
     return writeState();
 }
